Updates to get baseclasses to work (also added helpers.py for not base classes)
